test(api): split HTTP contract suites by resource
