+++ services/api/src/test/java/dev/dictum/api/PostsHttpContractTest.java
@@ -3,87 +3,14 @@ package dev.dictum.api;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import dev.dictum.api.generated.model.PostResponse;
-import dev.dictum.api.generated.model.SessionResponse;
-import dev.dictum.api.generated.model.SiteSettingsResponse;
-import dev.dictum.api.support.SessionHttpClient;
+import dev.dictum.api.support.InMemoryHttpContractSupport;
 import dev.dictum.api.support.SessionHttpClient.HttpMethod;
-import dev.dictum.api.support.TestAdminCredentials;
 import java.net.http.HttpResponse;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.web.server.LocalServerPort;
 import org.springframework.http.MediaType;
-import org.springframework.test.annotation.DirtiesContext;
 
-@SpringBootTest(
-    webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT,
-    properties = {
-      "dictum.content.repository=in-memory",
-      TestAdminCredentials.USERNAME_PROPERTY,
-      TestAdminCredentials.PASSWORD_PROPERTY
-    })
-@DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
-class HttpContractTest {
-
-  private static final String CONTENT_TYPE_HEADER = "content-type";
-  private static final String MERGE_PATCH_JSON = "application/merge-patch+json";
-  private static final String PARAMS_FIELD = "params";
-  private static final String POSTS_SEGMENT = "posts";
-  private static final String DICTUM_BEGINS_SLUG = "dictum-begins";
-  private static final String REMOTE_CONTROLS_LATER_SLUG = "remote-controls-later";
-  private static final String UNKNOWN_SLUG = "unknown-slug";
-  private static final String ADMIN_TAG = "admin";
-  private static final String API_TAG = "api";
-  private static final String SESSION_PATH = path("api", "v1", "session");
-  private static final String POSTS_PATH = path("api", "v1", POSTS_SEGMENT);
-  private static final String REMOTE_CONTROLS_LATER_PATH =
-      path("api", "v1", POSTS_SEGMENT, REMOTE_CONTROLS_LATER_SLUG);
-  private static final String SITE_SETTINGS_PATH = path("api", "v1", "settings", "site");
-  private static final String ADMIN_USERNAME = TestAdminCredentials.USERNAME;
-  private static final String ADMIN_PASSWORD = TestAdminCredentials.SECRET;
-
-  private final ObjectMapper objectMapper = new ObjectMapper().findAndRegisterModules();
-  private SessionHttpClient sessionHttpClient;
-
-  @LocalServerPort private int port;
-
-  @BeforeEach
-  void setUp() {
-    sessionHttpClient = new SessionHttpClient(baseUrl(), objectMapper);
-  }
-
-  @Test
-  void getSessionReturnsCurrentAuthenticatedSession() throws Exception {
-    HttpResponse<String> loginResponse =
-        sessionHttpClient.createSession(ADMIN_USERNAME, ADMIN_PASSWORD);
-
-    assertThat(loginResponse.statusCode()).isEqualTo(200);
-
-    HttpResponse<String> response = sessionHttpClient.get(SESSION_PATH);
-
-    assertThat(response.statusCode()).isEqualTo(200);
-
-    SessionResponse session = objectMapper.readValue(response.body(), SessionResponse.class);
-    assertThat(session.getUsername()).isEqualTo(ADMIN_USERNAME);
-    assertThat(session.getCsrfToken()).isNotBlank();
-  }
-
-  @Test
-  void createSessionRejectsInvalidCredentials() throws Exception {
-    HttpResponse<String> response =
-        sessionHttpClient.createSession(ADMIN_USERNAME, "wrong-password");
-
-    assertThat(response.statusCode()).isEqualTo(401);
-    assertThat(response.headers().firstValue(CONTENT_TYPE_HEADER))
-        .hasValueSatisfying(
-            value -> assertThat(value).contains(MediaType.APPLICATION_PROBLEM_JSON_VALUE));
-
-    JsonNode problem = objectMapper.readTree(response.body());
-    assertThat(problem.get("code").asText()).isEqualTo("auth.invalid_credentials");
-  }
+class PostsHttpContractTest extends InMemoryHttpContractSupport {
 
   @Test
   void protectedEndpointsRejectUnauthenticatedRequests() throws Exception {
@@ -397,44 +324,6 @@ class HttpContractTest {
   }
 
   @Test
-  void getSiteSettingsReturnsTheCurrentSiteValues() throws Exception {
-    HttpResponse<String> response =
-        sessionHttpClient.getAuthenticated(SITE_SETTINGS_PATH, ADMIN_USERNAME, ADMIN_PASSWORD);
-
-    assertThat(response.statusCode()).isEqualTo(200);
-
-    SiteSettingsResponse settings =
-        objectMapper.readValue(response.body(), SiteSettingsResponse.class);
-    assertThat(settings.getTitle()).isEqualTo("Dictum");
-    assertThat(settings.getSubtitle()).isEqualTo("A remotely steerable markdown blog kit.");
-  }
-
-  @Test
-  void updateSiteSettingsReturnsTheUpdatedRepresentation() throws Exception {
-    HttpResponse<String> response =
-        sessionHttpClient.requestAuthenticated(
-            HttpMethod.PATCH,
-            SITE_SETTINGS_PATH,
-            MERGE_PATCH_JSON,
-            """
-            {
-              "subtitle": "A modular markdown blog platform.",
-              "motd": "API-first contract work is underway."
-            }
-            """,
-            ADMIN_USERNAME,
-            ADMIN_PASSWORD);
-
-    assertThat(response.statusCode()).isEqualTo(200);
-
-    SiteSettingsResponse settings =
-        objectMapper.readValue(response.body(), SiteSettingsResponse.class);
-    assertThat(settings.getTitle()).isEqualTo("Dictum");
-    assertThat(settings.getSubtitle()).isEqualTo("A modular markdown blog platform.");
-    assertThat(settings.getMotd()).isEqualTo("API-first contract work is underway.");
-  }
-
-  @Test
   void unsafeRequestsRequireCsrfToken() throws Exception {
     sessionHttpClient.createSession(ADMIN_USERNAME, ADMIN_PASSWORD);
 
@@ -446,25 +335,5 @@ class HttpContractTest {
 
     JsonNode problem = objectMapper.readTree(response.body());
     assertThat(problem.get("code").asText()).isEqualTo("request.csrf_invalid");
-  }
-
-  @Test
-  void deleteSessionInvalidatesTheCurrentSession() throws Exception {
-    sessionHttpClient.createSession(ADMIN_USERNAME, ADMIN_PASSWORD);
-
-    HttpResponse<String> response = sessionHttpClient.delete(SESSION_PATH);
-
-    assertThat(response.statusCode()).isEqualTo(204);
-
-    HttpResponse<String> sessionResponse = sessionHttpClient.get(SESSION_PATH);
-    assertThat(sessionResponse.statusCode()).isEqualTo(401);
-  }
-
-  private String baseUrl() {
-    return "http://localhost:" + port;
-  }
-
-  private static String path(String... segments) {
-    return "/" + String.join("/", segments);
   }
 }
